requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cpu
torchvision==0.10.0
numpy==1.21.1
Pillow==8.3.1
gunicorn==20.1.0